% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mnre_dim_and_class_to_index}
\alias{mnre_dim_and_class_to_index}
\title{Index from dimension and class indicies}
\usage{
mnre_dim_and_class_to_index(i_dim, i_class, n_dim)
}
\arguments{
\item{i_dim}{integer}

\item{i_class}{integer}

\item{n_dim}{integer}
}
\value{
The integer index corresponding to the dimension and class indicies
}
\description{
Index from dimension and class indicies
}
\examples{
mnre_dim_and_class_to_index(1, 1, 1)
mnre_dim_and_class_to_index(2, 3, 5)
}
