% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mnre_expand_matrix}
\alias{mnre_expand_matrix}
\title{Expand matrix}
\usage{
mnre_expand_matrix(x1, k_class, direction)
}
\arguments{
\item{x1}{sparse matrix}

\item{k_class}{integer}

\item{direction}{integer}
}
\value{
Matrix expanded in the The integer index corresponding to the dimension and class indicies
}
\description{
Used to duplicate the random effects coefficient matrix before applying the left factor
}
