% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mnre_fit_sparse}
\alias{mnre_fit_sparse}
\title{Multinomial mixed effects fit}
\usage{
mnre_fit_sparse(fixed_effects, random_effects, y, theta_mat, Lind, beta_fixed,
  beta_random, verbose = 1L)
}
\arguments{
\item{x1}{sparse matrix}

\item{k_class}{integer}

\item{direction}{integer}
}
\value{
Matrix expanded in the The integer index corresponding to the dimension and class indicies
}
\description{
Used to duplicate the random effects coefficient matrix before applying the left factor
}
