% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mnre_make_covar}
\alias{mnre_make_covar}
\title{Make covariance matrix}
\usage{
mnre_make_covar(theta_mat, Lind, off_diagonal = 0)
}
\arguments{
\item{theta_mat}{numeric matrix}

\item{Lind}{unsigned integer matrix}

\item{off_diagonal}{numeric}
}
\value{
Covariance matrix for random effects
}
\description{
Make covariance matrix
}
\examples{
\dontrun{
theta_mat <- matrix(1:4, ncol=2)
Lind <- matrix(c(rep(1, 10), rep(2, 5)), ncol=1)
covar_mat <- mnre_make_covar(theta_mat, Lind)
}
}
