% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{nd_min_fun}
\alias{nd_min_fun}
\title{N-dimensional function generator}
\usage{
nd_min_fun(ev)
}
\arguments{
\item{ev}{list}
}
\value{
deviance function
}
\description{
N-dimensional function generator
}
\examples{
\dontrun{
ev = mnre_simulate_multinomial_data_factors(nfct=2, K_class = 2, nlev=50, nobs=20000)
nf <- nd_min_fun(ev)
nf(c(1,1))
}
}
